requirements.txt
Flask~=2.2.3
Flask-SQLAlchemy
psycopg2-binary
Jinja2~=3.1.4
MarkupSafe~=3.0.2
Werkzeug~=2.2.3
click~=8.1.7
blinker~=1.9.0
itsdangerous~=2.2.0

Flask-RESTx~=1.1.0
